Add HTML, CSS files and Images directory
This is my first commit for the Landing Page project.
Added the main files I think I will use for this project but
there could me more in a later commit.

I just learned about the Conventional Commits but not sure
if the initial commit has a <type>. That is why this one doesn't
adhere to the convention.

Note to self: Research more about the Conventional Commits in
order to avoid cementing bad habits for my commits in the future.