fix: style nav and hero with color, padding, flex
Added flexbox to the navbar and hero section as well as
some styling using the given instructions from The Odin Project
as well as some personal choices to adapt to the brand of the
mockup project.

Note to self: I definitely need to learn to be more patient and
prepare the project better and to not jump the gun just to code.